% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biophybreak_wrapper_functions.R
\name{continue.biophybreak}
\alias{continue.biophybreak}
\title{Continue sampling with biophybreak}
\usage{
continue.biophybreak(
  output,
  iter_max = 1e+05,
  iter_block = 10000,
  thin = 1,
  ESS_target = 200,
  save = TRUE,
  outdir = ".",
  seed = sample(2^31 - 1, 1)
)
}
\arguments{
\item{output}{The output from run.biophybreak that you wish to run longer}

\item{iter_max}{The maximum number of sampling iterations (including those already taken)}

\item{iter_block}{How frequently to check for sufficient effective sample size}

\item{thin}{Amount of thinning for MCMC samples}

\item{ESS_target}{Target effective sample size for all sampled parameters. Sampling will stop if this target is reached.}

\item{save}{Whether the MCMC output should be saved to a file}

\item{outdir}{The location at which the MCMC output should be saved}
}
\value{
A list containing the MCMC output, the input list, the effective sample sizes over time, 
the mixing status, the run name, and the computation time
}
\description{
Wrapper function to facilitate continue previous inference with biophybreak
}
