% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stat_functions.R
\name{descriptive.plots}
\alias{descriptive.plots}
\title{Plot descriptive/demographics statistics}
\usage{
descriptive.plots(
  df,
  max_groups = 10,
  save_plots = FALSE,
  run_set_title = "1",
  width = 9,
  height = 5
)
}
\arguments{
\item{df}{A data frame containing infection time distributions as in the output from prepare.HIV.data 
with find_infection_age_distributions = TRUE.}

\item{max_groups}{The maximum number of birth or suspected transmission locations to be plotting
before starting to group less common locations into "other"}

\item{save_plots}{Whether or not to save the plots as pdfs}

\item{run_set_title}{The first part of the file names if plots are saved}

\item{width}{Width of pdf outputs if plots are saved}

\item{height}{Height of pdf outputs if plots are saved}
}
\description{
Function to plot descriptive and demographic statistics of a dataset like gender and birth location distributions
}
