% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phybreak_analysis_functions.R
\name{group.cutoff}
\alias{group.cutoff}
\title{Pool smaller groups of individuals together}
\usage{
group.cutoff(df, group_name, max_groups)
}
\arguments{
\item{df}{A data frame containing infection time distributions as in the output from prepare.HIV.data}

\item{group_name}{The name of the column in df to consider}

\item{max_groups}{The maximum number of groups, including the "other" group}
}
\value{
A vector of the group that each individual belongs to with (potentially) the smallest groups pooled into "other"
}
\description{
Function to add cutoffs to grouping (pool smaller groups of individuals together in an "other" group)
}
