% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stat_functions.R
\name{infection.distributions.plot}
\alias{infection.distributions.plot}
\title{Plot densities of predicted infection times}
\usage{
infection.distributions.plot(df, sort_quant = 0.01, xmin_override = NULL)
}
\arguments{
\item{df}{A data frame containing infection time distributions as in the output from prepare.HIV.data 
with find_infection_age_distributions = TRUE.}

\item{sort_quant}{The quantile with which to chronologically sort clusters and individuals within clusters. 
Also determines the left boundary of the rectangle enclosing each cluster. 
For example, if set to 0.01, it uses the first percentile of the predicted infection times for sorting.}

\item{xmin_override}{If specified, this overrides the default method of determining the lower bound for plotting.}
}
\description{
Function to plot the individual and total densities of predicted infection times on one page (in two panels)
}
