% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phybreak_analysis_functions.R
\name{label.transmissions}
\alias{label.transmissions}
\title{Find transmission rates between individuals with certain labels}
\usage{
label.transmissions(
  MCMCstate,
  labels,
  infector.posterior.probabilities = phybreak.infector.posts(MCMCstate)$post_support
)
}
\arguments{
\item{MCMCstate}{The output from sample_phybreak}

\item{labels}{The labels for each individual}

\item{infector.posterior.probabilities}{The matrix of posterior probabilities of the infectors, 
as supplied by phybreak.infector.posts(MCMCstate)$post_support}
}
\description{
Function to find the expected number of transmission between and within individuals 
with some
}
