% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phybreak_analysis_functions.R
\name{phybreak.accuracy}
\alias{phybreak.accuracy}
\title{Find accuracy of phybreak inference}
\usage{
phybreak.accuracy(phybreak.true, MCMCstate, complete = TRUE)
}
\arguments{
\item{phybreak.true}{A phybreakdata object with the true transmission history}

\item{MCMCstate}{A phybreak object containing the information from the MCMC inference}

\item{complete}{Whether or not the MCMCstate inference was done with every individual in the true history.
Set to FALSE if there is at least one missing.}
}
\value{
A list containing the posterior support, accuracy, posterior increase factor, mean true posterior, 
a vector of whether each infector was inferred correctly, a vector of the posterior support the true infectors, 
the true infectors, and a dataframe with the posterior supports and true infectors.
}
\description{
Function to calculate the accuracy when comparing to the true simulated values
}
