% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phybreak_analysis_functions.R
\name{phybreak.infector.posts}
\alias{phybreak.infector.posts}
\title{Find phybreak posterior supports}
\usage{
phybreak.infector.posts(MCMCstate)
}
\arguments{
\item{MCMCstate}{A phybreak object containing the information from the MCMC inference}
}
\value{
A list containing a matrix and a dataframe of the posterior supports
}
\description{
Function to compute the posterior supports when the true infectors are unknown, 
i.e., when using real data.
}
