% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phybreak_analysis_functions.R
\name{phybreak.plot.posteriors}
\alias{phybreak.plot.posteriors}
\title{Barplot of posterior supports for infectors}
\usage{
phybreak.plot.posteriors(
  post_prob_df,
  treecolors = NULL,
  unsampled = NULL,
  ylab = "Posterior Support",
  ...
)
}
\arguments{
\item{post_prob_df}{A dataframe from either the phybreak.accuracy or phybreak.infector.posts function}

\item{treecolors}{Colors for each individual}

\item{unsampled}{Indices of individuals that were not considered sampled during inference if using simulated data}

\item{ylab}{Label of the y axis
... Other parameters that can be passed to ggplot::labs}
}
\description{
Function to create a barplot of the posterior supports for each infector
}
