% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phybreak_analysis_functions.R
\name{phybreak.plot.traces}
\alias{phybreak.plot.traces}
\title{Plot phybreak traces}
\usage{
phybreak.plot.traces(MCMCstate, main = "")
}
\arguments{
\item{MCMCstate}{A phybreak object containing the information from the MCMC inference}
}
\description{
Function to plot the MCMC traces
}
