% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phybreak_analysis_functions.R
\name{phybreak.plot.triple}
\alias{phybreak.plot.triple}
\title{Plot MPC tree, infection age distributions, and infector posterior supports}
\usage{
phybreak.plot.triple(MCMCstate, phybreak.true = NULL, angle = 90, ...)
}
\arguments{
\item{MCMCstate}{The output from sample_phybreak}

\item{phybreak.true}{The true transmission history (if using simulated data)}

\item{angle}{Angle of x-axis labels in the infector posterior density plot}

\item{...}{Other parameters that can be passed to ggplot::labs}
}
\description{
Function to plot the maximum parent credibility tree, infection age distributions, 
  and posterior supports for infectors in one page
}
