% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phybreak_analysis_functions.R
\name{phybreak.posterior}
\alias{phybreak.posterior}
\title{Find phybreak posterior probabilities}
\usage{
phybreak.posterior(MCMCstate)
}
\arguments{
\item{MCMCstate}{A phybreak object containing the information from the MCMC inference}
}
\value{
A list of the posterior probability and likelihoods
}
\description{
Function to compute the posterior probabilities for each MCMC iteration
}
