% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biophybreak_wrapper_functions.R
\name{prepare.biophybreak.data}
\alias{prepare.biophybreak.data}
\title{Create list of inputs for run.biophybreak}
\usage{
prepare.biophybreak.data(
  df,
  wh.model = "linear",
  wh.bottleneck = "wide",
  prior.wh.level.mean = 0.1,
  gen.type = "function",
  gen.density = stepfun(c(0, 0.5), c(0, 3, 1)),
  post.sam.trans.rate = 1,
  est.gen.mean = FALSE,
  use.pml = TRUE,
  mut.model = "pol",
  run.names = "automatic",
  name.prepend = "",
  overall.seed = sample(2^31 - 1, 1),
  ...
)
}
\arguments{
\item{df}{A dataframe as output from prepare.HIV.data}

\item{wh.model}{The model for within-host pathogen dynamics (effective pathogen population size = 
N*gE = actual population size * pathogen generation time), used to simulate coalescence events. Names and numbers are allowed.
Options are:
\enumerate{
  \item "single": effective size = 0, so coalescence occurs 'just before' transmission in the infector (complete bottleneck)
  \item "infinite": effective size = Inf, with complete bottleneck, so coalescence occurs 'just after' transmission in the infectee
  \item "linear": effective size at time t after infection = \code{wh.level + wh.slope * t} (complete or wide bottleneck; if complete, \code{wh.level = 0})
  \item "exponential": effective size at time t after infection = \code{wh.level * exp(wh.exponent * t)} (wide bottleneck)
  \item "constant": effective size = wh.level (wide bottleneck)
}}

\item{wh.bottleneck}{Whether the bottleneck should be complete or wide, which is only an option if \code{wh.model = "linear"} 
(in that case, \code{"auto"} defaults to \code{"complete"}).}

\item{prior.wh.level.mean}{Mean of the (gamma) prior distribution of \code{wh.level}}

\item{gen.type}{The type of input for the prior for the generation time. Can be 'parametric' (uses gamma distribution), 
'numeric' for use with a user-supplied density, or 'function' if the user supplied density is already a function.}

\item{gen.density}{Either a numeric distribution with x and y attributes, such as the output of the density() function
(if gen.type is 'numeric'), or a density function (if 'gen.type' is 'function')}

\item{post.sam.trans.rate}{The rate of transmission after sampling relative to transmission prior to sampling.
"1" means that there is it is not any less likely for an individual to infect another individual after being sampled.
"0" means that it is not possible for an individual to infect another after being sampled.
"0.1" means that an individual is 10 times less likely to infect another after being sampled.}

\item{est.gen.mean}{Whether to estimate the mean generation interval or keep it fixed.}

\item{use.pml}{Whether to use phangorn::pml to calculate sequences likelihoods.}

\item{run.names}{The names of the clusters to be used to saving files with run.biophybreak}

\item{name.prepend}{An identifier to add to the beginning of the file names for the output}

\item{overall.seed}{The seed used to generate the individual run seeds.}
}
\description{
Function to create a list of inputs to use with run.biophybreak
}
