% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phybreak_simulation_functions.R
\name{rtrans_tree}
\alias{rtrans_tree}
\title{Generate a transmission tree}
\usage{
rtrans_tree(
  nInd = 10,
  time.span = (nInd - 1) * 0.5,
  gen.dist.type = "gamma",
  gen.shape = 1,
  gen.mean = 100,
  gen.dist = NULL,
  infection.ages = 1.5,
  tr_het = "uniform_fixed",
  power.base = 2,
  sep = 0.1,
  post.sam.trans.rate = 1,
  seed = sample(2e+09, 1)
)
}
\arguments{
\item{nInd}{The number of individuals}

\item{time.span}{The amount of time between the first and last infection.}

\item{gen.dist.type}{The type of distribution for the generation function 
(either "gamma" or "function", a user supplied function)}

\item{gen.shape}{Shape of the gamma distribution for the generation function, if a gamma function is used.}

\item{gen.mean}{Mean of the gamma distribution for the generation function, if a gamma function is used.}

\item{gen.dist}{Generation distribution, if gen.dist.type = "function" is used.}

\item{infection.ages}{A number or vector for the amount of time between infection and sampling. Must be positive.}

\item{tr_het}{Type of transmission heterogeneity. Does not directly change transmission heterogeneity, 
but makes some transmission heterogeneities more or less likely. 
Can be "uniform_fixed" (no relative weight differences), "uniform_rnd" (relative weights are a uniform random variable), 
or "power" (weights follow a power law)}

\item{power.base}{Exponent base for the weights if tr_het = "power" is used.}

\item{sep}{The minimum amount of time (in years) between any two infections.}

\item{seed}{The pseudorandom number generator seed.}

\item{post.sam.trans.weight}{How much transmission after sampling is penalized. 
1 means no penalty. 0.1 Would be 10 times less likely.}
}
\value{
A transmission history with individual names, donors,infection times, and sampling times.
}
\description{
Function to generate a simple random transmission tree
}
