% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biophybreak_wrapper_functions.R
\name{run.biophybreak}
\alias{run.biophybreak}
\title{Create phybreak object and run burn-in and sampling}
\usage{
run.biophybreak(
  inputs,
  burn_iter = 10000,
  iter_max = 1e+05,
  iter_block = 10000,
  thin = 1,
  ESS_target = 200,
  save = TRUE,
  outdir = "",
  ...
)
}
\arguments{
\item{inputs}{A list containing data and parameters to the phybreak function that may change between runs}

\item{burn_iter}{The number of iterations for burn-in}

\item{iter_max}{The maximum number of sampling iterations}

\item{iter_block}{How frequently to check for sufficient effective sample size}

\item{thin}{Amount of thinning for MCMC samples}

\item{ESS_target}{Target effective sample size for all sampled parameters. Sampling will stop if this target is reached.}

\item{save}{Whether the MCMC output should be saved to a file}

\item{outdir}{The location at which the MCMC output should be saved}
}
\value{
A list containing the MCMC output, the input list, the effective sample sizes over time, 
the mixing status, the run name, and the computation time
}
\description{
Wrapper function to facilitate running inference with biophybreak
}
