% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biophybreak_wrapper_functions.R
\name{run.mbm}
\alias{run.mbm}
\title{Wrapper for mbm.predict}
\usage{
run.mbm(
  df,
  n.adapt = 1000,
  n.burn = 1000,
  n.iter = 1000,
  overall.seed = sample(2^31 - 1, 1)
)
}
\arguments{
\item{df}{A data frame containing biomarker information as in the output from prepare.HIV.data}

\item{n.adapt}{Number of iterations for model adaptation}

\item{n.burn}{Number of burn-in iterations}

\item{n.iter}{Number of sampling iterations}

\item{overall.seed}{RNG seed to use to generate other RNG seeds}
}
\value{
Lists for the infection age distributions both in terms of time before diagnosis and time before the first sequence
}
\description{
Wrapper function to call mbm.predict in parallel
}
