% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phybreak_simulation_functions.R
\name{sim.coal.phybreak}
\alias{sim.coal.phybreak}
\title{Simulate a phyreakdata transmission history}
\usage{
sim.coal.phybreak(
  tt,
  a = rep(0, dim(tt)[1]),
  b = rep(3, dim(tt)[1]),
  rhoD = rep(0, dim(tt)[1]),
  rhoR = rep(0, dim(tt)[1]),
  nSamples = rep(1, dim(tt)[1]),
  sample.times = NULL,
  gen.rate = 365/1.5,
  tr_window = rep(Inf, dim(tt)[1]),
  mut_model = "env",
  user_seqs = NULL,
  user_seq_hosts = NULL,
  user_seq_names = NULL,
  sim_index = 1,
  tree_loc = "",
  seed = sample(2e+09, 1)
)
}
\arguments{
\item{tt}{A transmission tree data frame. Must contain the elements "names", "donor", "t_inf", and "t_sam" (see below)
May be left blank if these are provided separately.}

\item{a}{A vector for the effective population size (inverse of coalescence rate) for each patient at the time of infection.
A value of 0 corresponds to a complete transmission bottleneck (only 1 lineage transmitted), 
with larger values indicating a wider transmission bottleneck (potentially multiple sampled lineages transmitted)
Typical values for a wide bottleneck may be around 20.}

\item{b}{A vector for the linear growth rate of the effective population size per generation (usually 1.5 days for HIV). 
Faster growth rates indicated faster growth of effective population size, 
resulting in lower coalescence rates near sampling (longer branch lengths).
Typical values are around 3.}

\item{rhoD}{A vector for the rates for REVERSE TIME migration from donor to recipient during an ongoing transmission window.
If this is non-zero, it indicates that it is possible for a recipient to transmit a lineage back to the individual who infected them.
This is typically set to 0. 
This value does not do anything for individuals who did not infect anyone else.}

\item{rhoR}{A vector for the rates of REVERSE TIME migration from recipient to donor during an ongoing transmission window.
If this is non-zero, it indicates that the recipient can be infected by the donor at multiple times during an ongoing transmission window.
If ongoing transmission is used, typical values are 0.1
This value does not do anything for the index case, as their infector is not simulated.}

\item{nSamples}{A vector for the number of sequences from each patient.
Can be left blank if sample.times is provided.
If neither nSamples nor sample.times are provided, it will default to one sequence per individual.
If nSamples is provided but not sample.times, every sequence will be assumed to come from t_sam for that individual}

\item{sample.times}{A list of vectors for the sample times for each individual.
For example, sample.times = list(c(2.2, 2.3, 2.7), c(2.5, 3.5)) 
would mean that the first patient has one sequence from time 2.2, one from 2.3, and one from 2.7
and the second patient has one sequence from 2.5 and one from 3.5.}

\item{gen.rate}{The generation rate in number of generations per year. 
Defaults to 365/1.5 (One generation is approximately 1.5 days)}

\item{tr_window}{A vector for the length of the transmission window a recipient has with their donor, 
used when rhoD or rhoR is non-zero, given in years after the recient is infected.
Defaults to infinity, indicating that the ongoing transmission windowis open indefinitely.
Has no effect if rhoD and rhoR are zero.
Has no effect for the index case.}

\item{mut_model}{The mutation model to be used if seq-gen is used for sequences, 
which can be "env" or "pol" for the HIV-1 envelope or polymerase genes}

\item{user_seqs}{User provided sequences to be used if seq-gen is not used, in order of sampling time}

\item{sim_index}{The name of the tree for saving files.}

\item{tree_loc}{The directory in which to output saved tree files.
Defaults to an empty character, resulting in saving in the current directory.}
}
\value{
A phybreakdata object representing the transmission history, sample times, phylogeny, and sequences
}
\description{
Function to simulate phylogenic tree and sequences in phybreakdata format from a transmission histoey
}
