% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phybreak_simulation_functions.R
\name{sim.coal.tree}
\alias{sim.coal.tree}
\title{Simulate a coalescence tree}
\usage{
sim.coal.tree(
  tt = NULL,
  names = NULL,
  donor = NULL,
  t_inf = NULL,
  t_sam = NULL,
  a = rep(0, dim(tt)[1]),
  b = rep(3, dim(tt)[1]),
  rhoD = rep(0, dim(tt)[1]),
  rhoR = rep(0, dim(tt)[1]),
  nSamples = NULL,
  sample.times = NULL,
  gen.rate = 365/1.5,
  tr_window = rep(Inf, dim(tt)[1]),
  file_path = "",
  tree_name = 1,
  tree_name_digits = nchar(as.character(tree_name)),
  save_tree = FALSE,
  plot_tree = FALSE,
  plot_colors = NULL,
  seed = sample(2e+09, 1)
)
}
\arguments{
\item{tt}{A transmission tree data frame. Must contain the elements "names", "donor", "t_inf", and "t_sam" (see below)
May be left blank if these are provided separately.}

\item{names}{A vector of the names of the patients}

\item{donor}{A vector for the infectors for each patient. 0 indicated the index case (the infector was not sampled) 
and all other values must be positive integers corresponding to each individual's infector.}

\item{t_inf}{The infection times of each of the patients in years, usually relative to the infection time of the index patient.}

\item{a}{A vector for the effective population size (inverse of coalescence rate) for each patient at the time of infection.
A value of 0 corresponds to a complete transmission bottleneck (only 1 lineage transmitted), 
with larger values indicating a wider transmission bottleneck (potentially multiple sampled lineages transmitted)
Typical values for a wide bottleneck may be around 20.}

\item{b}{A vector for the linear growth rate of the effective population size per generation (usually 1.5 days for HIV). 
Faster growth rates indicated faster growth of effective population size, 
resulting in lower coalescence rates near sampling (longer branch lengths).
Typical values are around 3.}

\item{rhoD}{A vector for the rates for REVERSE TIME migration from donor to recipient during an ongoing transmission window.
If this is non-zero, it indicates that it is possible for a recipient to transmit a lineage back to the individual who infected them.
This is typically set to 0. 
This value does not do anything for individuals who did not infect anyone else.}

\item{rhoR}{A vector for the rates of REVERSE TIME migration from recipient to donor during an ongoing transmission window.
If this is non-zero, it indicates that the recipient can be infected by the donor at multiple times during an ongoing transmission window.
If ongoing transmission is used, typical values are 0.1
This value does not do anything for the index case, as their infector is not simulated.}

\item{nSamples}{A vector for the number of sequences from each patient.
Can be left blank if sample.times is provided.
If neither nSamples nor sample.times are provided, it will default to one sequence per individual.
If nSamples is provided but not sample.times, every sequence will be assumed to come from t_sam for that individual}

\item{sample.times}{A list of vectors for the sample times for each individual.
For example, sample.times = list(c(2.2, 2.3, 2.7), c(2.5, 3.5)) 
would mean that the first patient has one sequence from time 2.2, one from 2.3, and one from 2.7
and the second patient has one sequence from 2.5 and one from 3.5.}

\item{gen.rate}{The generation rate in number of generations per year. 
Defaults to 365/1.5 (One generation is approximately 1.5 days)}

\item{tr_window}{A vector for the length of the transmission window a recipient has with their donor, 
used when rhoD or rhoR is non-zero, given in years after the recient is infected.
Defaults to infinity, indicating that the ongoing transmission windowis open indefinitely.
Has no effect if rhoD and rhoR are zero.
Has no effect for the index case.}

\item{file_path}{The directory in which to output saved tree files and plots.
Defaults to an empty character, resulting in saving in the current directory.}

\item{tree_name}{The name of the tree for saving files. 
Defaults to 1, resulting in the file "tree1.txt" (if save_tree == TRUE and tree_name_digits is 3)}

\item{tree_name_digits}{The number of characters in the tree name. 
Defaults to the number of characters in the given tree name, so the name with both defaults for name and digits 
would be "tree1.txt". If 3 is supplied, the name would be tree001.txt. This does not do anything if tree_name is a character.}

\item{save_tree}{if TRUE, it will save the newick tree along with the tip labels and states in a text file}

\item{plot_tree}{If TRUE, this will plot the coalescent tree the pdf file "treeplot(tree_name).pdf"}

\item{plot_colors}{The colors to use for the tip labels for each individual. 
If there are only two individuals, it defaults to red for teh donor and blue for the recipient.
Otherwise defaults to the r base graphics colors.}
}
\value{
a coalescent tree in the Newick format (with branch lengths) and tip labels and states.
}
\description{
Simulation of a coalescence tree given a transmission tree
}
\examples{
tree <- sim.coal.tree(names = c("D", "R"), donor = c(0,1), t_inf = c(0,2), t_sam = c(3,4),
                      a = rep(20, 2), b = rep(3, 2),
                      nSamples = rep(500, 2))
tree <- sim.coal.tree(names = c("D", "R"), donor = c(0,1), t_inf = c(0,2), t_sam = c(3,4),
                      rhoD = rep(0, 2), rhoR = rep(0.1, 2),
                      sample.times = list(c(rep(2.2, 200), rep(2.3, 100), rep(3, 200)), c(rep(2.5, 250), rep(4, 250))),
                      tr_window = rep(2, 2))
}
