% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phybreak_simulation_functions.R
\name{subsam.seqs}
\alias{subsam.seqs}
\title{Subsample phybreakdata}
\usage{
subsam.seqs(sim_data, nseq_sub)
}
\arguments{
\item{sim_data}{A phybreakdata object to subsample sequences from}

\item{nseq_sub}{A number or vector for the number of sequences to keep from each individual}
}
\value{
A subsampled phybreakdata object
}
\description{
Subsamples sequences from phybreakdata simulation data
}
