% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phybreak_analysis_functions.R
\name{mpcinfectors}
\alias{mpcinfectors}
\title{Find MPC infectors}
\usage{
mpcinfectors(MCMCstate)
}
\arguments{
\item{MCMCstate}{A phybreak object containing the information from the MCMC inference}
}
\value{
The names and indices of the MPC infectors for each individual
}
\description{
Function to find the maximum parent credibility infectors from a phybreak object
}
